Separated cat-task models into own file
