+++ categories/categories-model.js
@@ -6,9 +6,7 @@ module.exports = {
   findCategories,
   add,
   update,
-  remove,
-  assignCategory,
-  findTasks
+  remove
 };
 
 function find() {
@@ -44,18 +42,4 @@ function remove(id) {
   return db('categories')
       .where({id})
       .del()
-}
-
-function assignCategory(combo) {
-  return db('task-categories')
-    .insert(combo)
-}
-
-// (id) below is the category id
-function findTasks(id) {
-  return db('task-categories as tc')
-    .join('tasks as t', 't.id', 'tc.task_id')
-    .join('categories', 'categories.id', 'tc.category_id')
-    .select('t.name', 't.status', 't.date', 't.start_time', 't.end_time', 't.task_icon', 't.timeLeft', 't.initialNotify', 't.notifyOn')
-    .where({category_id: id})
 }
